Introduce the command model
Introduce the command model, being the Wallet as an Aggregate. This
aggregate is in charge of handling the create, withdraw and deposit
command and deciding whether that operation might be taken. If so, a
corresponding created, withdrawn or deposited event will be published.
Decision making logic should be included upon the withdraw command, as
sufficient funds should be available. If this is not the case, an
InsufficientFundsException should be thrown

#step2